Add files via upload
initial conversion from at1 version to this version. im not sure if i was supposed to start a new repository but it was the only way it worked for me. this upload, is the first one, where i got all the information from the AT1, into the "strata" format. it is all still out of place though all the info is there