Add files via upload
in this commit, all the links were updated to be operational, each image/pdf was resized to be the correct size, each of the headings/subheadings and layout issues were fixed. The Navigation styling was also moved from each HTML file, to the main.css file. This commit has smoothed out most issues